+++ orix/quaternion/misorientation.py
@@ -31,6 +31,7 @@ import numpy as np
 from scipy.spatial.transform import Rotation as SciPyRotation
 from tqdm import tqdm
 
+from orix._utils.deprecation import deprecated
 from orix.quaternion.orientation_region import OrientationRegion
 from orix.quaternion.rotation import Rotation
 from orix.quaternion.symmetry import C1, Symmetry, _get_unique_symmetry_elements
@@ -340,6 +341,7 @@ class Misorientation(Rotation):
 
         return self.__class__(equivalent).flatten()
 
+    @deprecated(since="0.14", removal="0.15", alternative="reduce")
     def map_into_symmetry_reduced_zone(self, verbose: bool = False) -> Misorientation:
         """Return equivalent transformations which have the smallest
         angle of rotation as a new misorientation.
@@ -365,22 +367,52 @@ class Misorientation(Rotation):
         [[-0.7071  0.7071  0.      0.    ]
         [ 0.      1.      0.      0.    ]]
         """
+        return self.reduce(verbose)
+
+    def reduce(self, verbose: bool = False) -> Misorientation:
+        """Return equivalent transformations which have the smallest
+        angle of rotation as a new misorientation.
+
+        Parameters
+        ----------
+        verbose
+            Whether to print a progressbar. Default is ``False``.
+
+        Returns
+        -------
+        M
+            A new misorientation object with the assigned symmetry.
+
+        Examples
+        --------
+        >>> from orix.quaternion.symmetry import C4, C2
+        >>> data = np.array([[0.5, 0.5, 0.5, 0.5], [0, 1, 0, 0]])
+        >>> M = Misorientation(data)
+        >>> M.symmetry = (C4, C2)
+        >>> M.reduce()
+        Misorientation (2,) 4, 2
+        [[-0.7071  0.7071  0.      0.    ]
+        [ 0.      1.      0.      0.    ]]
+        """
         Gl, Gr = self._symmetry
+        fz = OrientationRegion.from_symmetry(Gl, Gr)
         symmetry_pairs = iproduct(Gl, Gr)
         if verbose:
             symmetry_pairs = tqdm(symmetry_pairs, total=Gl.size * Gr.size)
 
-        orientation_region = OrientationRegion.from_symmetry(Gl, Gr)
-        o_inside = self.__class__.identity(self.shape)
+        # There is one and only one combination of `symmetry_pairs` that
+        # moves an arbitrary rotation into fz. Apply the combinations
+        # iteratively to the outside quaternions until all are inside.
+        reduced = self.__class__.identity(self.shape)
         outside = np.ones(self.shape, dtype=bool)
         for gl, gr in symmetry_pairs:
             o_transformed = gl * self[outside] * gr
-            o_inside[outside] = o_transformed
-            outside = ~(o_inside < orientation_region)
+            reduced[outside] = o_transformed
+            outside = ~(reduced < fz)
             if not np.any(outside):
                 break
-        o_inside._symmetry = (Gl, Gr)
-        return o_inside
+        reduced._symmetry = (Gl, Gr)
+        return reduced
 
     def scatter(
         self,